requirements.txt
safetensors==0.4.0
Pillow==9.0.1
